pyproject.toml
[project]
name = "sil-website"
version = "0.1.0"
description = "Semantic Infrastructure Lab public website - clean Python web app demonstrating SIL architecture"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Scott", email = "scott@semanticinfrastructurelab.org"}
]
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "httpx>=0.25.0",
    "markdown>=3.5.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "reveal-cli>=0.11.0",
    "httpx-mock>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"
extend-select = ["I"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
